Sprint 8 solution time and duration

+++ test/FileBackedTaskManagerTest.java
@@ -3,29 +3,42 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import java.io.File;
 import java.io.IOException;
+import java.io.UncheckedIOException;
 import java.nio.file.Files;
 import static org.junit.jupiter.api.Assertions.*;
 
-class FileBackedTaskManagerTest {
-    private FileBackedTaskManager fileManager;
+class FileBackedTaskManagerTest extends TaskManagerTest<FileBackedTaskManager>{
     private File tempFile;
+
     @BeforeEach
-    void setUp() throws IOException {
-        this.tempFile = Files.createTempFile("task_manager_test", ".csv").toFile();
-        this.fileManager = new FileBackedTaskManager(tempFile);
+    @Override
+    void setUp()  {
+        try {
+            this.tempFile = Files.createTempFile("task_manager_test", ".csv").toFile();
+        } catch (IOException e){
+            throw new UncheckedIOException("Could not create temp file for test", e);
+        }
+        super.setUp();
     }
     @AfterEach
     void deleteTemFile() throws IOException {
-        Files.deleteIfExists(this.tempFile.toPath());
+        if (tempFile != null){
+            Files.deleteIfExists(this.tempFile.toPath());
+        }
+    }
+
+    @Override
+    protected FileBackedTaskManager createTaskManager() {
+        return new FileBackedTaskManager(tempFile);
     }
     @Test
     void correctLoadingAndUnloading() {
-        Task task = new Task("Тесты", "Не хватает времени написать тесты(");
-        fileManager.addTask(task);
+        Task task = new Task("Тесты", "Не хватает времени написать тесты");
+        this.taskManager.addTask(task);
         Epic epic = new Epic("Завершить 7 спринт","важное дело, чтобы начать восьмой");
-        fileManager.addEpic(epic);
+        this.taskManager.addEpic(epic);
         Subtask subtask = new Subtask(epic.getId(),"доделать метод записи","Весьма непростое занятие");
-        fileManager.addSubtask(subtask);
+        this.taskManager.addSubtask(subtask);
         FileBackedTaskManager loadingFile = FileBackedTaskManager.loadFromFile(tempFile);
         assertNotNull(loadingFile, "Менеджер не должен быть null");
         int tasksNumber = loadingFile.getTasks().size() + loadingFile.getEpics().size()
